refactored home view
