+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart' as DotEnv;
 import 'package:softshop_mobile/app/router.dart';
 import 'package:softshop_mobile/styles/theme/theme.dart' as _theme;
-import 'package:softshop_mobile/views/home/home_view_model.dart';
+import 'package:softshop_mobile/views/home/home_base_view.dart';
 import 'package:softshop_mobile/views/splash/splash_view.dart';
 import 'core/utils/exports.dart';
