--- lib/views/home/home_model.dart
@@ -1,3 +1,0 @@
-import 'package:softshop_mobile/core/utils/exports.dart';
-
-class HomeModel extends IndexTrackingViewModel {}
